+++ backend/.env
@@ -3,5 +3,3 @@ DB_PORT="5432"
 DB_USER="postgres"
 DB_NAME="prolingo"
 DB_PWD="password"
-
-
