--- backend/api/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path("notes/", views.NoteListCreate.as_view(), name="note-list-create"),
-    path("notes/delete/<int:pk>/", views.NoteDelete.as_view(), name="delete-note"),
-]
