+++ backend/notes/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+class NotesConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "notes"
