+++ backend/notes/models.py
@@ -1,7 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
 
-
 class Note(models.Model):
     title = models.CharField(max_length=100)
     content = models.TextField()
@@ -10,12 +9,3 @@ class Note(models.Model):
 
     def __str__(self):
         return self.title
-
-
-
-class Item(models.Model):
-    name = models.CharField(max_length=100)
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return self.name
