+++ backend/notes/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path("create", views.NoteListCreate.as_view(), name="note-list-create"),
+    path("delete/<int:pk>/", views.NoteDelete.as_view(), name="delete-note"),
+]
