+++ backend/requirements.txt
@@ -1,4 +1,3 @@
-Django==5.2.5
 asgiref
 Django
 django-cors-headers
