+++ backend/user/serializers.py
@@ -1,8 +1,5 @@
 from django.contrib.auth.models import User
 from rest_framework import serializers
-from .models import Item, Note
-
-
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
@@ -13,14 +10,3 @@ class UserSerializer(serializers.ModelSerializer):
     def create(self, validated_data):
         user = User.objects.create_user(**validated_data)
         return user
-    
-class NoteSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Note
-        fields = ['id', 'title', 'content', 'created_at', 'author']
-        extra_kwargs = {'author': {'read_only': True}}
-
-class ItemSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Item
-        fields = '__all__'
