+++ backend/user/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+from . import views
+
+urlpatterns = [
+    path("register/", views.CreateUserView.as_view(), name="register"),
+]
