+++ backend/user/views.py
@@ -1,0 +1,10 @@
+from django.shortcuts import render
+from django.contrib.auth.models import User
+from rest_framework import viewsets, generics
+from rest_framework.permissions import IsAuthenticated, AllowAny
+from .serializers import UserSerializer
+
+class CreateUserView(generics.CreateAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+    permission_classes = [AllowAny]
