+++ web/.gitignore
@@ -1,4 +1,3 @@
-# Logs
 logs
 *.log
 npm-debug.log*
@@ -7,12 +6,15 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
+__pycache__
+.venv
+env
+
 node_modules
 dist
 dist-ssr
 *.local
 
-# Editor directories and files
 .vscode/*
 !.vscode/extensions.json
 .idea
